Time: 7 ms (100%), Space: 45 MB (97.13%) - LeetHub

+++ 3195-separate-black-and-white-balls/3195-separate-black-and-white-balls.java
@@ -11,7 +11,7 @@ class Solution
             count1++;
             if(ch[i] == '1')
             {
-                count2 += count1;
+                count2 += (long)count1;
                
             }
